first changes to ResidueTemplateLibraries. They now handle position as well. We should figure out how to not add more parameters.
